break out prompt building
