+++ packages/server/src/sdk/workspace/ai/agents/index.ts
@@ -1,1 +1,2 @@
 export * from "./crud"
+export * from "./utils"
